telegram_bot/pyproject.toml
[tool.poetry]
name = "bot template"
version = "0.1.0"
description = "My template for creating bots with Django+DRF+Aiogram+AioHttp"
authors = ["https://github.com/itisnotyourenv"]

[tool.poetry.dependencies]
python = "^3.8"
aiogram = "^2.17.1"
django-environ = "^0.8.1"
APScheduler = "^3.8.1"


[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
